+++ src/core/trulens/experimental/otel_tracing/core/instrument.py
@@ -4,6 +4,7 @@ from typing import Callable, List, Optional
 import uuid
 
 from opentelemetry import trace
+from opentelemetry.baggage import get_baggage
 from opentelemetry.baggage import remove_baggage
 from opentelemetry.baggage import set_baggage
 import opentelemetry.context as context_api
@@ -110,17 +111,24 @@ class OTELRecordingContext:
     context api keeps track of what is changed in the context, and used to undo the changes.
     """
 
+    context_keys_added: List[str] = []
+    """
+    Keys added to the OTEL context.
+    """
+
     def __init__(self, *, app: core_app.App, run_name: str, input_id: str):
         self.app = app
         self.run_name = run_name
         self.input_id = input_id
         self.tokens = []
+        self.context_keys_added = []
         self.span_context = None
 
     # Calling set_baggage does not actually add the baggage to the current context, but returns a new one
     # To avoid issues with remembering to add/remove the baggage, we attach it to the runtime context.
     def attach_to_context(self, key: str, value: object):
         self.tokens.append(context_api.attach(set_baggage(key, value)))
+        self.context_keys_added.append(key)
 
     # For use as a context manager.
     def __enter__(self):
@@ -130,7 +138,10 @@ class OTELRecordingContext:
 
         tracer = trace.get_tracer_provider().get_tracer(TRULENS_SERVICE_NAME)
 
-        self.attach_to_context(SpanAttributes.RECORD_ID, otel_record_id)
+        # There might be nested records - in those scenarios use the outer one.
+        if not get_baggage(SpanAttributes.RECORD_ID):
+            self.attach_to_context(SpanAttributes.RECORD_ID, otel_record_id)
+
         self.attach_to_context(SpanAttributes.APP_NAME, self.app.app_name)
         self.attach_to_context(SpanAttributes.APP_VERSION, self.app.app_version)
 
@@ -171,15 +182,11 @@ class OTELRecordingContext:
             # TODO[SNOW-1854360]: Add in feature function spans.
             self.span_context.__exit__(exc_type, exc_value, exc_tb)
 
-        remove_baggage(SpanAttributes.RECORD_ID)
-        remove_baggage(SpanAttributes.APP_NAME)
-        remove_baggage(SpanAttributes.APP_VERSION)
-
-        # Safe to remove baggage keys even if we did not set them
-        remove_baggage(SpanAttributes.RUN_NAME)
-        remove_baggage(SpanAttributes.INPUT_ID)
-
         logger.debug("Exiting the OTEL app context.")
+
+        # Clearing the context / baggage added.
+        while self.context_keys_added:
+            remove_baggage(self.context_keys_added.pop())
 
         while self.tokens:
             # Clearing the context once we're done with this root span.
